requirement.txt
Flask
python-dotenv
Flask-Cors
Flask-JWT-Extended
pymongo
werkzeug
